CI: unset LD_LIBRARY_PATH
During teardown our BATS test call gpdeletesystem which stops the
postgres processes. This was failing with:

 Error: unable to import module: /usr/local/greenplum-db-6.20.2+dev.3.g24b949d/lib/libpq.so.5: symbol gss_acquire_cred_from, version gssapi_krb5_2_MIT not defined in file libgssapi_krb5.so.2 with link time reference

This is most likely because the LD_LIBRARY_PATH is already set in the
parent environment and when sourcing the target greenplum_path.sh this
also sets the LD_LIBRARY_PATH causing it to be polluted with both the
source and target cluster values. Thus, ensure LD_LIBRARY_PATH is unset
before sourcing the target greenplum_path.sh.

Co-authored-by: Edward Espino <eespino@vmware.com>
Co-authored-by: Kevin Yeap <kyeap@vmware.com>